pptagent/requirements.txt
openai>=1.50.0
python-pptx>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.30.0
tavily-python>=0.5.0
